Remove message checks
Not sure if this was an update or original functionality but it turns
out the request is rejected if permissions are incorrect or the message
is illegal: before the page was checked to make sure there is a new
message and that it matches. This is unreliable anyways because
messages can end up different.