Removed stream, added settings
Working with streams in this case just doesn't work, the object has to
be checked as to not duplicate players and streams are more complex
than a simple object in general. Although I originally did it for
memory, when you really think about it the names of players in a group
can't take that much memory, right? It can be re-added if there is
every trouble.