Fix for required argument error
Running into kind of a paradox here. The only way to check if the first
argument is an options object is if the object contains required
arguments of the function. So if it doesn't it will be assumed the
object is the first argument and not options. This means that even if I
send in an options object but miss a required argument, it will assume
it is not an options object and when the error occurs it won't be able
to figure out which argument is missing, although it will not if this
assumption had to be made. I could check if at least 1 required
argument exists to determine if it is an options object or not but this
is getting way too complicated, for now users will have to refer to the
documentation to figure out which argument they missed.