Patch
ROBLOX removed the login API I was using without warning. The reason I
was using it was because it returned login info but I suppose I should
not have depended on such an obscure API. It has been switched to the
given api on api.roblox.com which should be a lot more reliable;
however, I have been forced to make a major API change to the login
function because it can no longer return user info. I've decided not to
push to 3.0.0 (as would be required by semver following an API change
that is not backwards compatible) but I will not do so because this is
such a major patch and realistically not many people will depend on the
return value of login.

My apologies for the inconvenience but there was very little I could do
when ROBLOX doesn't inform anyone of an API change on their website.