Update for reliability
ROBLOX's APIs for join requests are ridiculously unstable. The search
url especially will frequently not work for no apparent reason. This
fixes a bug that would match users incorrectly and also tries to
circumvent the search by getting the request id directly from the group
admin page. If it does have to search eventually it tries its best to
appease ROBLOX and the function will now detect if that specifically
doesn't work.