Access by IP
The problem all along was that ROBLOX only created the temporary URL on
the remote address you were accessing when you requested when you
loaded the group admin page. This is obviously stupid and buggy as hell
on their end, when you create a separate URL it should not be isolated
on that specific server, because of DNS updates it has the potential to
break even for people using a regular browser. Note that because
accessing IP directly doesn't work well with certificates it will not
verify the authenticity of the responses when accessing the group admin
page and join request search page.