Fix
Requests were not being made as "asynchronous" posts which only load
the changed information (in this case the new page of players). This is
not only much more stressing on the network but also creates a bug that
if the specified group has a clan it will constantly store each page of
the clan every time.

Note to self that when using "__ASYNCPOST" User-Agent must be set to
something.