Fix ID case change
ROBLOX changed the "ID" field to "Id", completely disregarding how such a major API change could affect their users (as usual). In an effort to preserve the current major API for this module and to not have to edit a bunch of other files, the "Id" field will automatically be converted to "ID" when roles are returned.