Fix for exile.js (#76)
Opening a new pull request because the previous one was accidentally closed by me.
So, it seems like the current exile.js doesn't work when it gets an error. It always returns "Cannot read property length of undefined".
I added a JSON.parse() to parse the string, and it seems to work now.